Ajout de la méthode pour récupérer les informations des films via TmDb. Ajout tests unitaires pour cette méthode.

+++ Zeus.Test/Codes/ShowsAndMoviesManagerTest.cs
@@ -26,19 +26,29 @@ namespace Zeus.Test.Codes
 
             // Suppression des fichiers de configuration créé.
             DirectoryInfo directoryApp = new DirectoryInfo(AppContext.BaseDirectory + "/save");
-            directoryApp.Delete(true);
+            if (directoryApp.Exists)
+            {
+                directoryApp.Delete(true);
+            }
 
             DirectoryInfo directoryConfig = new DirectoryInfo(AppContext.BaseDirectory + "/config");
-            directoryConfig.Delete(true);
+            if (directoryConfig.Exists)
+            {
+                directoryConfig.Delete(true);
+            }
+
+            DirectoryInfo directorySave = new DirectoryInfo(AppContext.BaseDirectory + "/save");
+            if (directorySave.Exists)
+            {
+                directorySave.Delete(true);
+            }
         }
 
-
-
-
+        #region Method GetListMoviesLocal
+        
         [TestMethod]
         public void RecuperationDeFilms_When_MovieFolderIsEmpty_Then_NoFilm()
         {
-
             #region Act
 
             ShowsAndMoviesManager manager = new ShowsAndMoviesManager();
@@ -55,11 +65,9 @@ namespace Zeus.Test.Codes
 
 
         [TestMethod]
-        public void RecuperationDeFilms_When_MovieFolderHas2Films_Then_2Films()
+        public void RecuperationDeFilms_When_MovieFolderHas3Films_Then_3Films()
         {
             #region Arrange
-
-            
 
             if (!Directory.Exists(PATH_MOVIE))
             {
@@ -97,6 +105,57 @@ namespace Zeus.Test.Codes
             #endregion
         }
 
+        #endregion
+
+        #region Method GetMovies
+
+        [TestMethod]
+        public void RecuperationDeFilmsAvecTmDb_When_MovieFolderHas3Films_Then_3Films()
+        {
+            #region Arrange
+
+            if (!Directory.Exists(PATH_MOVIE))
+            {
+                Directory.CreateDirectory(PATH_MOVIE);
+            }
+
+            // Création de faux films.
+            DirectoryInfo directoryMovie = new DirectoryInfo(PATH_MOVIE);
+
+            var fsAmercian = File.Create(directoryMovie.FullName + @"\An.American.Haunting.2005.TRUEFRENCH.DVDRip.XviD.Mp3-Tetine.avi");
+            var movie2 = File.Create(directoryMovie.FullName + @"\paddington-2-french-bluray-1080p-2018.mkv");
+            var movie3 = File.Create(directoryMovie.FullName + @"\mazinger-z-french-bluray-1080p-2018.mp4");
+            var other1 = File.Create(directoryMovie.FullName + @"\fichierWeb.html");
+            var other2 = File.Create(directoryMovie.FullName + @"\readme.txt");
+
+            fsAmercian.Dispose();
+            movie2.Dispose();
+            movie3.Dispose();
+            other2.Dispose();
+            other1.Dispose();
+
+            #endregion
+
+            #region Act
+
+            ShowsAndMoviesManager manager = new ShowsAndMoviesManager();
+            var result = manager.GetMovies().Result;
+
+            #endregion
+
+            #region Assert
+
+            Assert.AreEqual(3, result.Count(), "Doit contenir 3 films.");
+
+            DirectoryInfo directorySave = new DirectoryInfo(AppContext.BaseDirectory + "/save");
+            var files = directorySave.GetFiles();
+
+            Assert.AreEqual(2, files.Count(), "Il doit y avoir 2 fichiers de sauvegarde");
+
+            #endregion
+        }
+
+        #endregion
 
     }
 }
